+++ CHANGELOG.md
@@ -130,6 +130,7 @@
 ### Fixes
 
 * Fix [#4017](https://github.com/buefy/buefy/issues/4017) `Tooltip` - AbortController is not defined in SSR mode (Nuxt)
+* Fix [#4018](https://github.com/buefy/buefy/issues/4018) Browser tab froze when `Field` with both `grouped` and `horizontal` props `true` got a validation error. However, there is another issue that validation errors cannot be reset once they are set if `grouped` and `horizontal` are combined.
 
 ## 0.9.28